+++ mrb_doc/core/input/mouse.rb
@@ -1,11 +1,3 @@
-# Returns the movement of the mouse wheel since last frame.
-# @return [Float]
-def get_mouse_wheel_move
-  # mrb_get_mouse_wheel_move
-  # src/mruby_integration/core/input/mouse.cpp
-  0.5
-end
-
 # Sets the mouse cursor.
 # @param cursor [Integer]
 # @return [nil]