+++ src/mruby_integration/core/input/mouse.cpp
@@ -7,12 +7,6 @@
 #include "ruby/core/input/mouse.hpp"
 
 auto
-mrb_get_mouse_wheel_move(mrb_state* mrb, mrb_value) -> mrb_value
-{
-  return mrb_float_value(mrb, GetMouseWheelMove());
-}
-
-auto
 mrb_set_mouse_cursor(mrb_state* mrb, mrb_value) -> mrb_value
 {
   mrb_int cursor;
@@ -26,11 +20,6 @@ mrb_set_mouse_cursor(mrb_state* mrb, mrb_value) -> mrb_value
 void
 append_core_input_mouse(mrb_state* mrb)
 {
-  mrb_define_method(mrb,
-                    mrb->kernel_module,
-                    "get_mouse_wheel_move",
-                    mrb_get_mouse_wheel_move,
-                    MRB_ARGS_NONE());
   mrb_define_method(mrb,
                     mrb->kernel_module,
                     "set_mouse_cursor",